Rename all the projects to conform to the new agreed naming scheme for caching projects.
