+++ src/Polly.Caching.Distributed.NetStandard11.Specs/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+﻿using System.Reflection;
+using Xunit;
+
+[assembly: AssemblyTitle("Polly.Caching.Distributed.NetStandard11.Specs")]
+[assembly: CollectionBehavior(DisableTestParallelization = true)]
