--- src/Polly.Caching.IDistributedCache.NetStandard11.Specs/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿using System.Reflection;
-using Xunit;
-
-[assembly: AssemblyTitle("Polly.Caching.IDistributedCache.NetStandard11.Specs")]
-[assembly: CollectionBehavior(DisableTestParallelization = true)]
