--- src/Polly.Caching.IDistributedCache.NetStandard11/Polly.Caching.IDistributedCache.NetStandard11.csproj
@@ -1,37 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <AssemblyName>Polly.Caching.IDistributedCache</AssemblyName>
-    <AssemblyOriginatorKeyFile>..\Polly.Caching.IDistributedCache.Net45\Polly.Caching.IDistributedCache.snk</AssemblyOriginatorKeyFile>
-    <DefineConstants>TRACE;PORTABLE</DefineConstants>
-    <DefaultLanguage>en-US</DefaultLanguage>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <RootNamespace>Polly.Caching.IDistributedCache</RootNamespace>
-    <TargetFramework>netstandard1.1</TargetFramework>
-    <NetStandardImplicitPackageVersion>1.6.1</NetStandardImplicitPackageVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugType>full</DebugType>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <WarningsAsErrors />
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType>pdbonly</DebugType>
-    <Optimize>true</Optimize>
-  </PropertyGroup>
-  <ItemGroup>
-    <NuSpecFile Include="$(SolutionDir)Polly.Caching.IDistributedCache.nuspec">
-      <Visible>False</Visible>
-    </NuSpecFile>
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="..\GlobalAssemblyInfo.cs">
-      <Link>Properties\GlobalAssemblyInfo.cs</Link>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="1.1.2" />
-    <PackageReference Include="Polly" Version="5.4.0" />
-  </ItemGroup>
-  <Import Project="..\Polly.Caching.IDistributedCache.Shared\Polly.Caching.IDistributedCache.Shared.projitems" Label="Shared" />
-</Project>
