--- src/Polly.Caching.IDistributedCache.Shared/TtlHelper.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Extensions.Caching.Distributed;
-
-namespace Polly.Caching.IDistributedCache
-{
-    internal static class InternalTtlHelper
-    {
-        internal static DistributedCacheEntryOptions ToDistributedCacheEntryOptions(this Ttl ttl)
-        {
-            return ttl.SlidingExpiration
-                ? new DistributedCacheEntryOptions { SlidingExpiration = ttl.Timespan }
-                : new DistributedCacheEntryOptions { AbsoluteExpirationRelativeToNow = ttl.Timespan };
-        }
-    }
-}
