--- src/Polly.Caching.IDistributedCache.SharedSpecs/Polly.Caching.IDistributedCache.SharedSpecs.projitems
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>8d537c37-a4a2-4be1-84bc-34f1afb51ca8</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Polly.Caching.IDistributedCache.Specs</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)CacheProviderHelperTests.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NetStandardIDistributedCacheByteArrayProviderSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NetStandardIDistributedCacheStringProviderSpecs.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)TtlStrategyHelperTests.cs" />
-  </ItemGroup>
-</Project>
