--- src/Polly.Caching.IDistributedCache.sln
@@ -1,59 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.16
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{79FE9DBE-1155-4F78-A41F-FE95DEFA19DD}"
-	ProjectSection(SolutionItems) = preProject
-		GlobalAssemblyInfo.cs = GlobalAssemblyInfo.cs
-		Polly.nuspec = Polly.nuspec
-	EndProjectSection
-EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Polly.Caching.IDistributedCache.Shared", "Polly.Caching.IDistributedCache.Shared\Polly.Caching.IDistributedCache.Shared.shproj", "{23FA87F9-C77D-4C67-A0B0-2901DE51B3FF}"
-EndProject
-Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Polly.Caching.IDistributedCache.SharedSpecs", "Polly.Caching.IDistributedCache.SharedSpecs\Polly.Caching.IDistributedCache.SharedSpecs.shproj", "{8D537C37-A4A2-4BE1-84BC-34F1AFB51CA8}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.IDistributedCache.NetStandard11.Specs", "Polly.Caching.IDistributedCache.NetStandard11.Specs\Polly.Caching.IDistributedCache.NetStandard11.Specs.csproj", "{4AB46D16-0806-4F51-8677-D849BA87CB05}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.IDistributedCache.NetStandard11", "Polly.Caching.IDistributedCache.NetStandard11\Polly.Caching.IDistributedCache.NetStandard11.csproj", "{5017174E-DAC5-4408-AB15-1F902F40C612}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.IDistributedCache.NetStandard20", "Polly.Caching.IDistributedCache.NetStandard20\Polly.Caching.IDistributedCache.NetStandard20.csproj", "{BEB48D73-7C49-49AD-A53E-EBE7897833D7}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Polly.Caching.IDistributedCache.NetStandard20.Specs", "Polly.Caching.IDistributedCache.NetStandard20.Specs\Polly.Caching.IDistributedCache.NetStandard20.Specs.csproj", "{2D38A483-3D6D-4B50-BE51-DBFFB960EF69}"
-EndProject
-Global
-	GlobalSection(SharedMSBuildProjectFiles) = preSolution
-		Polly.Caching.IDistributedCache.Shared\Polly.Caching.IDistributedCache.Shared.projitems*{23fa87f9-c77d-4c67-a0b0-2901de51b3ff}*SharedItemsImports = 13
-		Polly.Caching.IDistributedCache.SharedSpecs\Polly.Caching.IDistributedCache.SharedSpecs.projitems*{4ab46d16-0806-4f51-8677-d849ba87cb05}*SharedItemsImports = 4
-		Polly.Caching.IDistributedCache.Shared\Polly.Caching.IDistributedCache.Shared.projitems*{5017174e-dac5-4408-ab15-1f902f40c612}*SharedItemsImports = 4
-		Polly.Caching.IDistributedCache.SharedSpecs\Polly.Caching.IDistributedCache.SharedSpecs.projitems*{8d537c37-a4a2-4be1-84bc-34f1afb51ca8}*SharedItemsImports = 13
-	EndGlobalSection
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4AB46D16-0806-4F51-8677-D849BA87CB05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4AB46D16-0806-4F51-8677-D849BA87CB05}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4AB46D16-0806-4F51-8677-D849BA87CB05}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4AB46D16-0806-4F51-8677-D849BA87CB05}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5017174E-DAC5-4408-AB15-1F902F40C612}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5017174E-DAC5-4408-AB15-1F902F40C612}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5017174E-DAC5-4408-AB15-1F902F40C612}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5017174E-DAC5-4408-AB15-1F902F40C612}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BEB48D73-7C49-49AD-A53E-EBE7897833D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BEB48D73-7C49-49AD-A53E-EBE7897833D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BEB48D73-7C49-49AD-A53E-EBE7897833D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BEB48D73-7C49-49AD-A53E-EBE7897833D7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2D38A483-3D6D-4B50-BE51-DBFFB960EF69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2D38A483-3D6D-4B50-BE51-DBFFB960EF69}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2D38A483-3D6D-4B50-BE51-DBFFB960EF69}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2D38A483-3D6D-4B50-BE51-DBFFB960EF69}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B79F0C1F-1732-4033-8D25-C26DAD98B10D}
-	EndGlobalSection
-EndGlobal
